requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
boto3>=1.34.0
sentence-transformers>=2.7.0
